move existing utilities to main app, remove appcore project in gradle
