+++ app/src/main/java/org/torproject/android/OrbotApp.kt
@@ -5,8 +5,8 @@ import android.content.res.Configuration
 import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.ProcessLifecycleOwner
-import org.torproject.android.core.Languages
-import org.torproject.android.core.LocaleHelper
+import org.torproject.android.util.Languages
+import org.torproject.android.util.LocaleHelper
 import org.torproject.android.service.util.Prefs
 
 import java.util.Locale
