+++ app/src/main/java/org/torproject/android/ui/v3onionservice/OnionServiceDeleteDialogFragment.kt
@@ -10,9 +10,8 @@ import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.DialogFragment
 
 import org.torproject.android.R
-import org.torproject.android.core.DiskUtils.recursivelyDeleteDirectory
+import org.torproject.android.util.DiskUtils.recursivelyDeleteDirectory
 import org.torproject.android.service.OrbotConstants
-import org.torproject.android.service.db.OnionServiceColumns
 
 import java.io.File
