+++ app/src/main/java/org/torproject/android/ui/v3onionservice/clientauth/ClientAuthActivity.java
@@ -19,7 +19,7 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
 import org.torproject.android.R;
-import org.torproject.android.core.DiskUtils;
+import org.torproject.android.util.DiskUtils;
 import org.torproject.android.ui.components.BaseActivity;
 import org.torproject.android.service.db.V3ClientAuthColumns;
 import org.torproject.android.ui.v3onionservice.V3BackupUtils;
