+++ app/src/main/java/org/torproject/android/util/ClipboardUtils.kt
@@ -1,4 +1,4 @@
-package org.torproject.android.core
+package org.torproject.android.util
 
 import android.content.ClipData
 import android.content.ClipboardManager
