+++ app/src/main/java/org/torproject/android/util/DiskUtils.kt
@@ -1,4 +1,4 @@
-package org.torproject.android.core
+package org.torproject.android.util
 
 import android.content.ContentResolver
 import android.content.Context
