+++ app/src/main/java/org/torproject/android/util/LocaleHelper.kt
@@ -1,4 +1,4 @@
-package org.torproject.android.core
+package org.torproject.android.util
 
 import android.content.Context
 import org.torproject.android.service.util.Prefs
