+++ settings.gradle
@@ -20,7 +20,6 @@ dependencyResolutionManagement {
 rootProject.name = "Orbot"
 include(
         ':app',
-        ':appcore',
         ':intentintegrator',
         ':OrbotLib',
         ':orbotservice'
